# perf: improve performance