--- a/src/components/shared/Description.tsx
+++ b/src/components/shared/Description.tsx
@@ -1,9 +1,8 @@
-import { Box } from '@stoplight/mosaic';
 import * as React from 'react';
 
 export const Description: React.FunctionComponent<{ value: string }> = ({ value }) => (
   // TODO (JJ): Add mosaic popover showing full description in MarkdownViewer
-  <Box title={value} textOverflow="truncate" w="full" color="muted">
+  <div className="sl-truncate sl-w-full sl-text-muted" title={value}>
     {value}
-  </Box>
+  </div>
 );