--- a/src/components/shared/Divider.tsx
+++ b/src/components/shared/Divider.tsx
@@ -1,5 +1,4 @@
 import { SchemaCombinerName } from '@stoplight/json-schema-tree';
-import { Box, Flex } from '@stoplight/mosaic';
 import * as React from 'react';
 
 import { COMBINER_PRETTY_NAMES } from '../../consts';
@@ -8,12 +7,10 @@ export const Divider: React.FunctionComponent<{ kind?: SchemaCombinerName; style
   kind,
   style,
 }) => (
-  <Flex align="center" w="full" pos="absolute" style={{ ...style, ...(kind && { top: -9 }), height: 1 }}>
+  <div className="sl-flex sl-text-center sl-w-full sl-absolute" style={{ ...style, ...(kind && { top: -9 }), height: 1 }}>
     {kind && (
-      <Box pr={2} ml={-6} textTransform="uppercase" fontSize="sm" color="muted">
-        {COMBINER_PRETTY_NAMES[kind]}
-      </Box>
+      <div className="sl-pr-2 sl--ml-6 sl-uppercase sl-text-sm sl-text-muted">{COMBINER_PRETTY_NAMES[kind]}</div>
     )}
-    {!kind && <Box flex={1} style={{ height: 1, backgroundColor: 'lightgray' }} />}
-  </Flex>
+    {!kind && <div className="sl-flex-1" style={{ height: 1, backgroundColor: 'lightgray' }} />}
+  </div>
 );