--- a/src/components/shared/Divider.tsx
+++ b/src/components/shared/Divider.tsx
@@ -7,7 +7,10 @@ export const Divider: React.FunctionComponent<{ kind?: SchemaCombinerName; style
   kind,
   style,
 }) => (
-  <div className="sl-flex sl-text-center sl-w-full sl-absolute" style={{ ...style, ...(kind && { top: -9 }), height: 1 }}>
+  <div
+    className="sl-flex sl-text-center sl-w-full sl-absolute"
+    style={{ ...style, ...(kind && { top: -9 }), height: 1 }}
+  >
     {kind && (
       <div className="sl-pr-2 sl--ml-6 sl-uppercase sl-text-sm sl-text-muted">{COMBINER_PRETTY_NAMES[kind]}</div>
     )}