--- a/src/components/shared/Validations.tsx
+++ b/src/components/shared/Validations.tsx
@@ -142,8 +142,7 @@ const NameValidations = ({ validations }: { validations: Dictionary<unknown> }) 
       .map(key => {
         return (
           <div className="sl-flex sl-flex-1 sl-my-2" key={key}>
-            <span              className="sl-px-1 sl-text-muted sl-font-mono sl-border sl-rounded-lg sl-text-sm sl-capitalize"
-            >
+            <span className="sl-px-1 sl-text-muted sl-font-mono sl-border sl-rounded-lg sl-text-sm sl-capitalize">
               {key}
             </span>
           </div>