… src/__tests__/__snapshots__/index.spec.tsx.snap
… src/__tests__/index.spec.tsx
… yarn.lock