Times in parenthesis are mounting times of the 2x-stress schema story in Storybook.
- Bumps @stoplightio/json-schema-merge-allof (1660ms -> 1590ms)
- Removes trivial mosaic components from the performance-critical rendering paths: `SchemaRow` and beyond (1590ms -> 850ms)

The problem with mosaic components is that for almost 0 gain (beyond type-checking of style-props), you render yet another element, and especially yet another `forwardRef`-ed element, both of which eat up performance as you scale.